pyproject.toml
[project]
name = "chat-with-dataframe"
version = "0.1.0"
description = "Chat with DataFrame."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["toolkit"]

[tool.setuptools.package-dir]
"email_assistant" = "toolkit"
